# docs/swagger.yaml
basePath: /
definitions:
  structure.Advertisement:
    description: 广告数据结构，包含广告的所有信息
    properties:
      action_url:
        description: "ActionUrl 点击广告跳转URL，可选\n\t@example\t\"https://example.com/promo\""
        type: string
      content:
        description: "Content 广告内容，可选\n\t@example\t\"春节期间全场商品8折优惠\""
        type: string
      created_at:
        description: "CreatedAt 广告创建时间，ISO 8601格式\n\t@example\t\"2024-01-15T10:00:00Z\""
        type: string
      end_time:
        description: "EndTime 广告结束时间，ISO 8601格式，必须晚于开始时间\n\t@example\t\"2024-02-15T23:59:59Z\""
        type: string
      image_url:
        description: "ImageURL 广告图片URL，必填\n\t@example\t\"https://example.com/advertisement.jpg\""
        type: string
      locale:
        description: "Locale 语言区域，遵循 RFC 5646 标准\n\t@example\t\"zh\""
        type: string
      start_time:
        description: "StartTime 广告开始时间，ISO 8601格式\n\t@example\t\"2024-02-01T00:00:00Z\""
        type: string
      title:
        description: "Title 广告标题，可选\n\t@example\t\"春节促销\""
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    提供推送通知相关的 API 接口，支持发送文本和图片通知，集成 OneSignal 推送服务，
    提供 WebSocket 实时通知功能，支持向所有用户广播通知，
    同时提供通知记录存储和管理功能，为应用提供完整的推送通知解决方案
  title: 推送通知服务
  version: "1.0"
paths:
  /advertisement/all:
    get:
      consumes:
      - application/json
      description: |-
        获取当前有效的所有广告列表，按创建时间降序排序。可以通过locale参数过滤特定语言的广告。

        功能说明：
        - 返回当前时间在有效期内的所有广告
        - 如果提供了locale参数，只返回匹配该语言的广告
        - 按创建时间降序排序（最新的在前）

        参数说明：
        - locale: 可选，语言区域过滤，支持的值：zh、en、zh-CN、zh-TW
        - zh: 简体中文
        - en: 英语
        - zh-CN: 中国大陆中文
        - zh-TW: 台湾中文

        使用示例：
        - /advertisement/all?locale=zh 获取所有中文广告
        - /advertisement/all?locale=en 获取所有英文广告
        - /advertisement/all 获取所有广告（不限语言）
      parameters:
      - description: 语言区域 (zh, en, zh-CN, zh-TW)
        in: query
        name: locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的locale参数
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有广告
      tags:
      - 广告管理
  /advertisement/create:
    post:
      consumes:
      - application/json
      description: |-
        创建一个新的广告并保存到数据库。广告包含标题、内容、图片URL、开始时间、结束时间、语言区域等信息。

        必填字段：
        - title: 广告标题，可为空字符串
        - content: 广告内容，可为空字符串
        - image_url: 广告图片URL，不能为空
        - start_time: 广告开始时间，ISO 8601格式
        - end_time: 广告结束时间，必须晚于开始时间
        - locale: 语言区域，遵循 RFC 5646 标准，例如 zh、en、zh-CN、zh-TW
        - action_url: 点击广告跳转URL（可选）

        注意事项：
        - 广告将在开始时间和结束时间之间有效
        - locale参数必须符合RFC 5646标准
        - image_url为必填字段
      parameters:
      - description: 广告信息
        in: body
        name: advertisement
        required: true
        schema:
          $ref: '#/definitions/structure.Advertisement'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建新广告
      tags:
      - 广告管理
  /advertisement/delete/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        根据广告ID从数据库中永久删除指定的广告记录。此操作不可恢复，请谨慎使用。

        参数说明：
        - id: 广告的唯一标识符（MongoDB ObjectID），必须存在于数据库中

        删除成功后，该广告将不再对用户可见，且无法恢复。
        如果指定的ID不存在，将返回404错误。
      parameters:
      - description: 广告 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 广告不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除广告
      tags:
      - 广告管理
  /advertisement/latest:
    get:
      consumes:
      - application/json
      description: |-
        获取当前有效的、最新的广告信息。可以通过locale参数过滤特定语言的广告。

        功能说明：
        - 返回当前时间在有效期内的最新广告
        - 如果提供了locale参数，只返回匹配该语言的广告
        - 如果没有找到广告，返回空数据

        参数说明：
        - locale: 可选，语言区域过滤，支持的值：zh、en、zh-CN、zh-TW
        - zh: 简体中文
        - en: 英语
        - zh-CN: 中国大陆中文
        - zh-TW: 台湾中文

        使用示例：
        - /advertisement/latest?locale=zh 获取最新的中文广告
        - /advertisement/latest?locale=en 获取最新的英文广告
        - /advertisement/latest 获取最新广告（不限语言）
      parameters:
      - description: 语言区域 (zh, en, zh-CN, zh-TW)
        in: query
        name: locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的locale参数
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取最新广告
      tags:
      - 广告管理
  /advertisement/update/{id}:
    put:
      consumes:
      - application/json
      description: |-
        根据广告ID更新广告信息。可以部分更新广告字段，只更新提供的字段。

        参数说明：
        - id: 广告的唯一标识符（MongoDB ObjectID），必须存在于数据库中
        - updateData: 要更新的字段键值对，支持部分更新

        可更新字段：
        - title: 广告标题
        - content: 广告内容
        - image_url: 广告图片URL
        - start_time: 广告开始时间
        - end_time: 广告结束时间
        - locale: 语言区域
        - action_url: 点击广告跳转URL

        注意事项：
        - 不能更新_id和id字段
        - 只更新提供的字段，未提供的字段保持不变
        - 时间格式必须符合ISO 8601标准
      parameters:
      - description: 广告 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新字段键值对
        in: body
        name: updateData
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 广告不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新广告
      tags:
      - 广告管理
...
swagger: "2.0"